pyproject.toml
[project]
name = "trading_agent"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "hyperliquid-python-sdk>=0.20.0,<0.21.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "web3>=7.14.0,<8.0.0",
    "aiohttp>=3.13.1,<4.0.0",
    "openai>=2.5.0,<3.0.0",
    "requests>=2.32.5,<3.0.0",
    "rich>=14.2.0,<15.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.rye]
src = ["src"]   